Highlight of selected text
This pastebin service contains lots of logs (mostly ClickHouse), and it
is useful to highlight selected text in the whole logs (query_id,
logger name, function name).

This is simpler implementation of [highlight-within-textarea](https://github.com/lonekorean/highlight-within-textarea) without jQuery.

_Note: this implementation is hackish enough, since I don't know either of CSS/HTML/JavaScript,
but want this feature since I'm looking lots of logs in this service_

Refs: https://codersblock.com/blog/highlight-text-inside-a-textarea